requirements.txt
fastapi==0.115.4
requests==2.32.3
pycryptodome==3.21.0
python-dotenv==1.0.1
gunicorn==23.0.0
uvicorn==0.32.0
sqlalchemy==2.0.36
pydantic==2.9.2
pyodbc==5.2.0
langchain==0.3.4
python-multipart==0.0.12
aiofiles==24.1.0
azure-identity==1.19.0
langchain_openai==0.2.4
azure-storage-blob==12.23.1
azure-cognitiveservices-speech==1.41.1
